requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Single-cell analysis
scanpy>=1.9.0
anndata>=0.9.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3

# Data download
GEOparse>=2.0.3
requests>=2.28.0

# Utilities
tqdm>=4.65.0
scikit-learn>=1.3.0

# Gene annotation
mygene>=3.2.2

# Optional for inferCNV (run in R separately)
# rpy2>=3.5.0  # if you want to call R from Python
